refactor: moved aws cloud config package to infra package
