+++ src/main/java/com/depromeet/archive/infra/cloud/aws/S3Service.java
@@ -3,9 +3,9 @@ package com.depromeet.archive.infra.cloud.aws;
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.depromeet.archive.config.cloud.aws.AwsS3Property;
 import com.depromeet.archive.domain.archive.ArchiveImageService;
 import com.depromeet.archive.exception.infra.FileInvalidException;
+import com.depromeet.archive.infra.cloud.aws.config.AwsS3Property;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
