+++ src/main/java/com/depromeet/archive/infra/cloud/aws/config/AwsConfiguration.java
@@ -1,4 +1,4 @@
-package com.depromeet.archive.config.cloud.aws;
+package com.depromeet.archive.infra.cloud.aws.config;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
