+++ src/main/java/com/depromeet/archive/infra/cloud/aws/config/AwsS3Property.java
@@ -1,4 +1,4 @@
-package com.depromeet.archive.config.cloud.aws;
+package com.depromeet.archive.infra.cloud.aws.config;
 
 import lombok.Getter;
 import lombok.Setter;
